// src/apis/models/Library.type.ts
export interface Book {
  title: string;
  subtitle: string;
  image: string;
  url: string;
  isbn13: string;
}

export interface BookInformation {
  title: string;
  subtitle: string;
  authors: string;
  publisher: string;
  pages: string;
  rating: string;
  desc: string;
  price: string;
  image: string;
}